Some OSS fixes to include data and work with gcp urls better.
PiperOrigin-RevId: 460790244

+++ learned_optimization/filesystem.py
@@ -19,17 +19,26 @@ import glob as py_glob
 import os
 import shutil
 from typing import Sequence
+import tensorflow as tf
 
 
+
+def _path_on_gcp(path: str) -> bool:
+  prefixes = ["gs://"]
+  return any([path.startswith(p) for p in prefixes])
 
 
 def file_open(path: str, mode: str):
   """Open a file, returning a file object."""
+  if _path_on_gcp(path):
+    return tf.io.gfile.GFile(path, mode)
   return open(path, mode)
 
 
 def make_dirs(path: str):
   """Make directories for given path."""
+  if _path_on_gcp(path):
+    return tf.io.gfile.makedirs(path)
 
   if not os.path.exists(path):
     return os.makedirs(path)
@@ -37,28 +46,38 @@ def make_dirs(path: str):
 
 def copy(path: str, target: str):
   """Copy path to target."""
+  if _path_on_gcp(path):
+    return tf.io.gfile.copy(path, target, overwrite=True)
 
   return shutil.copy(path, target)
 
 
 def rename(path: str, target: str):
   """Copy path to target."""
+  if _path_on_gcp(path):
+    return tf.io.gfile.rename(path, target, overwrite=True)
 
   return shutil.move(path, target)
 
 
 def exists(path: str) -> bool:
   """Check if a file exists."""
+  if _path_on_gcp(path):
+    return tf.io.gfile.exists(path)
 
   return os.path.exists(path)
 
 
 def glob(pattern: str) -> Sequence[str]:
   """Glob the filesystem with given pattern."""
+  if _path_on_gcp(pattern):
+    return tf.io.gfile.glob(pattern)
 
   return py_glob.glob(pattern)
 
 
 def remove(path: str) -> bool:
   """Remove a file."""
+  if _path_on_gcp(path):
+    return tf.io.gfile.remove(path)
   return shutil.rmtree(path)
